**Lesson: Models, Stores and Reducers**
We defined a model interface, we now need to implement it. We'll cover the basic concepts Model stores. A model store is a thread-safe way to managing state in an MVI application. The concept is borrowed from state container libraries like Flux and Redux. We will see how to use RxJava to build our own ModelState Store.